Created separated TaskManager.java class and moves methods from Main.java to it.
